Update: Review comment: moving css to common file
