+++ packages/terra-core-docs/src/terra-dev-site/doc/toggle-button/example/ButtonAttrsToggleButton.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Button from 'terra-button';
 import ToggleButton from 'terra-toggle-button';
 import classNames from 'classnames/bind';
-import styles from './OpenCloseEventToggleButton.module.scss';
+import styles from './ToggleButtonCommon.module.scss';
 
 const cx = classNames.bind(styles);
