--- packages/terra-core-docs/src/terra-dev-site/doc/toggle-button/example/OpenCloseEventToggleButton.module.scss
@@ -1,12 +1,0 @@
-:local { 
-  .times-wrapper {
-    display: inline-block;
-  }
-
-  .toggle-button-content {
-    a:focus {
-      outline: 2px dashed #000;
-      outline-offset: 2px;
-    }
-  }
-}
