+++ packages/terra-core-docs/src/terra-dev-site/doc/toggle-button/example/ToggleButtonCommon.module.scss
@@ -12,4 +12,15 @@
       outline-offset: 2px;
     }
   }
+
+  .times-wrapper {
+    display: inline-block;
+  }
+
+  .toggle-button-content {
+    a:focus {
+      outline: 2px dashed #000;
+      outline-offset: 2px;
+    }
+  }
 }
